LayoutTests/imported/w3c/web-platform-tests/webcodecs/videoFrame-canvasImageSource-expected.txt



Harness Error (TIMEOUT), message = null

TIMEOUT <video> and VideoFrame constructed VideoFrame Test timed out
PASS CSSImageValue constructed VideoFrame
PASS Wait for onload event to get access to image data
FAIL Image element constructed VideoFrame assert_equals: top left corner expected "Yellow" but got "#ff"
FAIL SVGImageElement constructed VideoFrame assert_equals: top left corner expected "Yellow" but got "#ff"
PASS Canvas element constructed VideoFrame
PASS Copy of canvas element constructed VideoFrame
